+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [1.0.1](https://github.com/privixy/nexora/compare/v1.0.0...v1.0.1) (2026-07-20)
+
+### Bug Fixes
+
+* handle imported connection passwords ([abc5e31](https://github.com/privixy/nexora/commit/abc5e31f73e96fb0d1c50f043d8bf61003496433))
+* update banner image size for better visibility ([c7698a1](https://github.com/privixy/nexora/commit/c7698a11da59485eb8f34764978f2673c1b7a8a7))
+
 # Changelog
 
 ## [0.15.0](#0150) (2026-07-17)